+++ packages/compass-ui/src/components/Modal/index.ts
@@ -1,2 +1,2 @@
 export { default as Modal } from './Modal';
-export type { ModalProps, ModalSize } from './Modal';
+export type { ModalBodyPadding, ModalProps, ModalSize } from './Modal';